backend/requirements.txt
asgiref==3.8.1
Django==4.2.19
django-filter==25.1
djangorestframework==3.15.2
djoser==2.3.1
flake8
gunicorn==20.1.0
psycopg2-binary==2.9.3
Pillow==11.1.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
